members/admin: sort age in waiting list only by year
Previously, the age column was sorted according to the birth date. This typically
causes singleton grouping upon multi sort with age as the primary sorting key. A
typical use case is sorting by age and application date.

Now the age column is sorted only by the numerical year value, which causes
the groups to be bigger.