ci: add pre-commit hooks (#12)
Pre-commit hooks can be used locally by installing pre-commit:
```
$ pip install pre-commit
```
and running the command directly or installing the hook to be run before
every commit:
```
$ pre-commit
$ pre-commit install
$ git commit
```

The re-commit action runs the same tests in a pipeline. Here, all files
altered by the PR or all added files in a commit are checked (when
pushing to main).