feat: add command to create test data (#199)
This command can be used to fill the database with some initial test
data. The command can be useful for both the staging deployment workflow
and local development. All members are created with email address under a configurable domain.

If the container is started with environment variable
`POPULATE_TEST_DATA=true`, the command is automatically run in the
entrypoint script after all migrations.